feat: Added getReferencedBy function
Fix isReferencedBy in same time.
